feat(shell): remove `update-index --assume-unchanged` abbrs
`git update-index --assume-unchanged` is intended for filesystems
without an efficient `lstat` implementation to cause Git not to do the
`st_mtime` information check to see if the file contents have changed
from the version recorded in the index file. **Not** for ignoring
changes to tracked files.

See [Using "Assume Unchanged" Bit] and [Notes] in the [`git
update-index` reference].

[Using "Assume Unchanged" Bit]:
https://git-scm.com/docs/git-update-index#_using_assume_unchanged_bit
[Notes]: https://git-scm.com/docs/git-update-index#_notes
[`git update-index` reference]:
https://git-scm.com/docs/git-update-index